Refactor: Extract identical code into a function
